Add OLPC XO-1.75 laptop support
This is a Marvell MMP2-based ARM laptop powered by OpenFirwmare. The
firmware loads \boot\olpc.fth on a first active partition. It has to be
a FAT or Ext* filesystem, with more recent features, such as extents,
not being supported.

This works well with the default layout with /boot being a separate ext2
partition, with a minor uglyness of olpc.fth being in /boot/boot/.